Update card abilities and refactor status effect control
Added abilities for 'Death' and 'TheHangedMan' cards in 'PlayerAttackMixin' and refactored effect control in 'LivingEntityMixin'. Status effects inflicted by 'TheHighPriestess' card only apply to the 'PlayerEntity'. The exceptions are handled to maintain game stability. The interfaces for player management have been extended with emperor, temperance and hierophant methods, in addition to inventory management and mana methods.