Implement 'The Fool' card abilities and attributes
This commit adds the ability and effects of 'The Fool' card in the CardManager class. Methods are added to create a bounding box around the player, apply effects within this box, and activate these abilities. Interfaces for fool cancellation are added in PlayerManagement interface and accordingly updated in CustomPlayerMixin. Finally, conditional checks have been integrated to enable or disable other cards based on 'The Fool' effect status.